Move context folder outside app folder
